requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.10.0
chardet==3.0.4
discord.py==1.6.0
idna==2.10
markdownify==0.10.3
multidict==4.7.6
python-dotenv==0.19.2
six==1.16.0
soupsieve==2.3.1
typing-extensions==4.0.1
yarl==1.5.1
